Rev.1 Выполнил задание по ТЗ, добавлено кеширование

+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/commerce/interaction/client/shopping/cart/ShoppingCartOperations.java
@@ -1,10 +1,5 @@
 package ru.yandex.practicum.commerce.interaction.client.shopping.cart;
 
-import jakarta.validation.Valid;
-import jakarta.validation.constraints.NotEmpty;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Positive;
-import org.springframework.web.bind.annotation.*;
 import ru.yandex.practicum.commerce.interaction.client.shopping.cart.exception.*;
 import ru.yandex.practicum.commerce.interaction.dto.shopping.cart.ShoppingCartDto;
 
@@ -16,7 +11,7 @@ public interface ShoppingCartOperations {
 
     ShoppingCartDto getShoppingCart(String username);
 
-    ShoppingCartDto addProductToShoppingCart(String username, Map<UUID, Integer> products);
+    ShoppingCartDto addProductToShoppingCart(String username, Map<UUID, Long> products);
 
     void deactivateCurrentShoppingCart(String username);
